% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{.od_artifact_build_id}
\alias{.od_artifact_build_id}
\title{Resolve a build UI number to internal id for artifact calls}
\usage{
.od_artifact_build_id(build_number, conn, use_internal_id = FALSE)
}
\description{
Resolve a build UI number to internal id for artifact calls
}
\keyword{internal}
