% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{.od_artifact_path}
\alias{.od_artifact_path}
\title{Normalize an artifact path for OneDev artifact endpoints}
\usage{
.od_artifact_path(artifact_path = NULL, leading_slash = FALSE)
}
\description{
Normalize an artifact path for OneDev artifact endpoints
}
\keyword{internal}
