% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{.od_as_tibble_default}
\alias{.od_as_tibble_default}
\title{Default for tibble conversion of list endpoints}
\usage{
.od_as_tibble_default(as_tibble = NULL)
}
\description{
Default for tibble conversion of list endpoints
}
\keyword{internal}
