% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{.od_default_commit_fields}
\alias{.od_default_commit_fields}
\title{Default commit fields requested from repository endpoints}
\usage{
.od_default_commit_fields()
}
\description{
Default commit fields requested from repository endpoints
}
\keyword{internal}
