% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.od_derive_project_path}
\alias{.od_derive_project_path}
\title{Derive a OneDev project path from a git remote URL}
\usage{
.od_derive_project_path(repo_url)
}
\description{
Derive a OneDev project path from a git remote URL
}
\keyword{internal}
