% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{.od_infer_auth}
\alias{.od_infer_auth}
\title{Infer auth mode from connection fields}
\usage{
.od_infer_auth(auth = NULL, username = "", token = "", password = "")
}
\description{
Infer auth mode from connection fields
}
\keyword{internal}
