% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builds.R
\name{.od_read_i32be}
\alias{.od_read_i32be}
\title{Read a signed big-endian 32-bit integer from raw bytes}
\usage{
.od_read_i32be(bytes, pos)
}
\description{
Read a signed big-endian 32-bit integer from raw bytes
}
\keyword{internal}
