% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{.od_repo_commit_query}
\alias{.od_repo_commit_query}
\title{Attach repeated \code{field} query params for commit endpoints}
\usage{
.od_repo_commit_query(query = NULL, count = NULL, fields = NULL)
}
\description{
Attach repeated \code{field} query params for commit endpoints
}
\keyword{internal}
