% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{.od_repo_path_encode}
\alias{.od_repo_path_encode}
\title{Encode path segments for repository file/directory URLs}
\usage{
.od_repo_path_encode(parts)
}
\description{
Encode path segments for repository file/directory URLs
}
\keyword{internal}
