% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.od_strip_hash}
\alias{.od_strip_hash}
\title{Strip a leading \verb{#} from a UI issue/build number}
\usage{
.od_strip_hash(x)
}
\description{
Strip a leading \verb{#} from a UI issue/build number
}
\keyword{internal}
