% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_requests.R
\name{od_approve_pull_request}
\alias{od_approve_pull_request}
\title{Approve a pull request}
\usage{
od_approve_pull_request(pull_request_number, conn = NULL)
}
\arguments{
\item{pull_request_number}{UI number.}

\item{conn}{Connection list.}
}
\value{
Parsed API response (may be \code{NULL}).
}
\description{
Approve a pull request
}
