% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_requests.R
\name{od_discard_pull_request}
\alias{od_discard_pull_request}
\title{Discard a pull request}
\usage{
od_discard_pull_request(pull_request_number, conn = NULL)
}
\arguments{
\item{pull_request_number}{UI number.}

\item{conn}{Connection list.}
}
\value{
Parsed API response (may be \code{NULL}).
}
\description{
Discard a pull request
}
\examples{
\dontrun{
od_discard_pull_request(1)
}
}
\seealso{
Other pull requests:
\code{\link[=od_add_pull_request_comment]{od_add_pull_request_comment()}},
\code{\link[=od_approve_pull_request]{od_approve_pull_request()}},
\code{\link[=od_create_pull_request]{od_create_pull_request()}},
\code{\link[=od_get_pull_request]{od_get_pull_request()}},
\code{\link[=od_get_pull_request_comments]{od_get_pull_request_comments()}},
\code{\link[=od_get_pull_request_reviews]{od_get_pull_request_reviews()}},
\code{\link[=od_merge_pull_request]{od_merge_pull_request()}},
\code{\link[=od_query_pull_requests]{od_query_pull_requests()}},
\code{\link[=od_request_pull_request_changes]{od_request_pull_request_changes()}},
\code{\link[=od_resolve_pull_request_id]{od_resolve_pull_request_id()}}
}
\concept{pull requests}
