% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{od_get_connection}
\alias{od_get_connection}
\title{Get the package-default OneDev connection}
\usage{
od_get_connection()
}
\value{
The connection registered with \code{\link[=od_set_connection]{od_set_connection()}}, or \code{NULL}.
}
\description{
Get the package-default OneDev connection
}
