% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{od_get_me}
\alias{od_get_me}
\title{Get the authenticated user (\verb{/users/me})}
\usage{
od_get_me(conn = NULL)
}
\arguments{
\item{conn}{Connection list from \code{\link[=od_get_config]{od_get_config()}} / \code{\link[=od_connection]{od_connection()}}.}
}
\value{
Parsed user object (list).
}
\description{
Get the authenticated user (\verb{/users/me})
}
