% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{od_get_pack}
\alias{od_get_pack}
\title{Get a package by id}
\usage{
od_get_pack(pack_id, conn = NULL)
}
\arguments{
\item{pack_id}{Package id.}

\item{conn}{Connection list from \code{\link[=od_get_config]{od_get_config()}} / \code{\link[=od_connection]{od_connection()}}.}
}
\value{
Parsed package object (list).
}
\description{
Get a package by id
}
\examples{
\dontrun{
od_get_pack(1)
}
}
\seealso{
Other packages:
\code{\link[=od_get_pack_blobs]{od_get_pack_blobs()}},
\code{\link[=od_get_pack_labels]{od_get_pack_labels()}},
\code{\link[=od_query_packages]{od_query_packages()}}
}
\concept{packages}
