% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{od_get_project}
\alias{od_get_project}
\title{Get a single project}
\usage{
od_get_project(project = NULL, conn = NULL)
}
\arguments{
\item{project}{Project path (e.g. \code{"group/my-project"}) or numeric id.
Defaults to the connection project.}

\item{conn}{Connection list from \code{\link[=od_get_config]{od_get_config()}} / \code{\link[=od_connection]{od_connection()}}.}
}
\value{
Parsed project object (list).
}
\description{
Get a single project
}
\examples{
\dontrun{
od_get_project()
}
}
\seealso{
Other projects:
\code{\link[=od_get_project_clone_url]{od_get_project_clone_url()}},
\code{\link[=od_query_projects]{od_query_projects()}},
\code{\link[=od_resolve_project_id]{od_resolve_project_id()}},
\code{\link[=od_resolve_project_path]{od_resolve_project_path()}}
}
\concept{projects}
