% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{od_get_tag}
\alias{od_get_tag}
\title{Get tag tip metadata}
\usage{
od_get_tag(tag, project = NULL, conn = NULL)
}
\arguments{
\item{tag}{Tag name.}

\item{project}{Project path or id; defaults to the connection project.}

\item{conn}{Connection list from \code{\link[=od_get_config]{od_get_config()}} / \code{\link[=od_connection]{od_connection()}}.}
}
\value{
List with \code{refName} and \code{commitHash}.
}
\description{
Get tag tip metadata
}
