% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onedevr-package.R
\docType{package}
\name{onedevr-package}
\alias{onedevr}
\alias{onedevr-package}
\title{onedevr: Access to the OneDev REST API}
\description{
R client for the OneDev project management REST API. Provides low-level request helpers and high-level convenience functions for issues, projects, and (future) builds. Designed for self-hosted OneDev instances. Modeled on gitlabr for the R API shape; OneDev conventions follow the tod CLI (https://github.com/theonedev/tod).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexseymer/onedevR}
  \item Report bugs at \url{https://github.com/alexseymer/onedevR/issues}
}

}
\author{
\strong{Maintainer}: Alex Seymer \email{alex@seymer.at}

Authors:
\itemize{
  \item Alex Seymer \email{alex@seymer.at}
}

}
\keyword{internal}
