% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builds.R
\name{od_get_build}
\alias{od_get_build}
\title{Get a single build by UI number}
\usage{
od_get_build(build_number, conn = NULL, use_internal_id = FALSE)
}
\arguments{
\item{build_number}{UI number (\code{100} or \code{"#100"}).}

\item{conn}{Connection list from \code{\link[=od_get_config]{od_get_config()}} / \code{\link[=od_connection]{od_connection()}}.}

\item{use_internal_id}{If \code{TRUE}, treat \code{build_number} as the internal REST
id (debugging only).}
}
\value{
Parsed build object (list).
}
\description{
Get a single build by UI number
}
\examples{
\dontrun{
od_get_build(100)
}
}
\seealso{
Other builds:
\code{\link[=od_cancel_job]{od_cancel_job()}},
\code{\link[=od_download_build_artifact]{od_download_build_artifact()}},
\code{\link[=od_get_build_log]{od_get_build_log()}},
\code{\link[=od_get_build_params]{od_get_build_params()}},
\code{\link[=od_list_build_artifacts]{od_list_build_artifacts()}},
\code{\link[=od_query_builds]{od_query_builds()}},
\code{\link[=od_rebuild_job]{od_rebuild_job()}},
\code{\link[=od_resolve_build_id]{od_resolve_build_id()}},
\code{\link[=od_run_job]{od_run_job()}}
}
\concept{builds}
