% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{od_get_commit}
\alias{od_get_commit}
\title{Get a single commit}
\usage{
od_get_commit(
  commit_hash,
  fields = .od_default_commit_fields(),
  project = NULL,
  conn = NULL
)
}
\arguments{
\item{commit_hash}{Commit SHA.}

\item{fields}{Character vector of fields to populate.}

\item{project}{Project path or id; defaults to the connection project.}

\item{conn}{Connection list from \code{\link[=od_get_config]{od_get_config()}} / \code{\link[=od_connection]{od_connection()}}.}
}
\value{
Parsed commit object (list).
}
\description{
Get a single commit
}
\examples{
\dontrun{
od_get_commit("abcdef")
}
}
\seealso{
Other repository:
\code{\link[=od_get_branch]{od_get_branch()}},
\code{\link[=od_get_default_branch]{od_get_default_branch()}},
\code{\link[=od_get_file]{od_get_file()}},
\code{\link[=od_get_file_text]{od_get_file_text()}},
\code{\link[=od_get_tag]{od_get_tag()}},
\code{\link[=od_list_branches]{od_list_branches()}},
\code{\link[=od_list_tags]{od_list_tags()}},
\code{\link[=od_query_commits]{od_query_commits()}}
}
\concept{repository}
