% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{od_get_issue}
\alias{od_get_issue}
\title{Get a single issue by UI number}
\usage{
od_get_issue(issue_number, conn = NULL, use_internal_id = FALSE)
}
\arguments{
\item{issue_number}{UI number (\code{145} or \code{"#145"}).}

\item{conn}{Connection list from \code{\link[=od_get_config]{od_get_config()}} / \code{\link[=od_connection]{od_connection()}}.}

\item{use_internal_id}{If \code{TRUE}, treat \code{issue_number} as the internal REST
id (debugging only).}
}
\value{
Parsed issue object (list).
}
\description{
Get a single issue by UI number
}
\examples{
\dontrun{
od_get_issue(145)
}
}
\seealso{
Other issues:
\code{\link[=od_add_issue_comment]{od_add_issue_comment()}},
\code{\link[=od_add_issue_iterations]{od_add_issue_iterations()}},
\code{\link[=od_create_issue]{od_create_issue()}},
\code{\link[=od_get_issue_comments]{od_get_issue_comments()}},
\code{\link[=od_get_issue_fields]{od_get_issue_fields()}},
\code{\link[=od_issue_set_description]{od_issue_set_description()}},
\code{\link[=od_issue_set_fields]{od_issue_set_fields()}},
\code{\link[=od_issue_set_title]{od_issue_set_title()}},
\code{\link[=od_issue_transition_state]{od_issue_transition_state()}},
\code{\link[=od_list_iterations]{od_list_iterations()}},
\code{\link[=od_query_issues]{od_query_issues()}},
\code{\link[=od_resolve_issue_id]{od_resolve_issue_id()}}
}
\concept{issues}
