% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{od_get_me}
\alias{od_get_me}
\title{Get the authenticated user (\verb{/users/me})}
\usage{
od_get_me(conn = NULL)
}
\arguments{
\item{conn}{Connection list from \code{\link[=od_get_config]{od_get_config()}} / \code{\link[=od_connection]{od_connection()}}.}
}
\value{
Parsed user object (list).
}
\description{
Get the authenticated user (\verb{/users/me})
}
\examples{
\dontrun{
od_get_me()
}
}
\seealso{
Other users:
\code{\link[=od_get_user]{od_get_user()}},
\code{\link[=od_get_user_emails]{od_get_user_emails()}},
\code{\link[=od_query_users]{od_query_users()}},
\code{\link[=od_resolve_user_id]{od_resolve_user_id()}}
}
\concept{users}
