% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{od_get_tag}
\alias{od_get_tag}
\title{Get tag tip metadata}
\usage{
od_get_tag(tag, project = NULL, conn = NULL)
}
\arguments{
\item{tag}{Tag name.}

\item{project}{Project path or id; defaults to the connection project.}

\item{conn}{Connection list from \code{\link[=od_get_config]{od_get_config()}} / \code{\link[=od_connection]{od_connection()}}.}
}
\value{
List with \code{refName} and \code{commitHash}.
}
\description{
Get tag tip metadata
}
\examples{
\dontrun{
od_get_tag("v1.0.0")
}
}
\seealso{
Other repository:
\code{\link[=od_get_branch]{od_get_branch()}},
\code{\link[=od_get_commit]{od_get_commit()}},
\code{\link[=od_get_default_branch]{od_get_default_branch()}},
\code{\link[=od_get_file]{od_get_file()}},
\code{\link[=od_get_file_text]{od_get_file_text()}},
\code{\link[=od_list_branches]{od_list_branches()}},
\code{\link[=od_list_tags]{od_list_tags()}},
\code{\link[=od_query_commits]{od_query_commits()}}
}
\concept{repository}
