% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{od_issue_set_description}
\alias{od_issue_set_description}
\title{Set an issue description}
\usage{
od_issue_set_description(issue_number, description, conn = NULL)
}
\arguments{
\item{issue_number}{UI number.}

\item{description}{New description (Markdown).}

\item{conn}{Connection list.}
}
\value{
Parsed API response.
}
\description{
Set an issue description
}
