% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{od_issue_set_fields}
\alias{od_issue_set_fields}
\title{Set custom issue fields}
\usage{
od_issue_set_fields(issue_number, fields, conn = NULL)
}
\arguments{
\item{issue_number}{UI number.}

\item{fields}{Named list of field values (installation-specific schema).}

\item{conn}{Connection list.}
}
\value{
Parsed API response.
}
\description{
Set custom issue fields
}
