% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{od_issue_set_title}
\alias{od_issue_set_title}
\title{Set an issue title}
\usage{
od_issue_set_title(issue_number, title, conn = NULL)
}
\arguments{
\item{issue_number}{UI number.}

\item{title}{New title.}

\item{conn}{Connection list.}
}
\value{
Parsed API response.
}
\description{
Set an issue title
}
