% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{od_issue_set_title}
\alias{od_issue_set_title}
\title{Set an issue title}
\usage{
od_issue_set_title(issue_number, title, conn = NULL)
}
\arguments{
\item{issue_number}{UI number.}

\item{title}{New title.}

\item{conn}{Connection list.}
}
\value{
Parsed API response.
}
\description{
Set an issue title
}
\examples{
\dontrun{
od_issue_set_title(145, "New title")
}
}
\seealso{
Other issues:
\code{\link[=od_add_issue_comment]{od_add_issue_comment()}},
\code{\link[=od_add_issue_iterations]{od_add_issue_iterations()}},
\code{\link[=od_create_issue]{od_create_issue()}},
\code{\link[=od_get_issue]{od_get_issue()}},
\code{\link[=od_get_issue_comments]{od_get_issue_comments()}},
\code{\link[=od_get_issue_fields]{od_get_issue_fields()}},
\code{\link[=od_issue_set_description]{od_issue_set_description()}},
\code{\link[=od_issue_set_fields]{od_issue_set_fields()}},
\code{\link[=od_issue_transition_state]{od_issue_transition_state()}},
\code{\link[=od_list_iterations]{od_list_iterations()}},
\code{\link[=od_query_issues]{od_query_issues()}},
\code{\link[=od_resolve_issue_id]{od_resolve_issue_id()}}
}
\concept{issues}
