% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{od_list_branches}
\alias{od_list_branches}
\title{List branches in a project}
\usage{
od_list_branches(project = NULL, conn = NULL)
}
\arguments{
\item{project}{Project path or id; defaults to the connection project.}

\item{conn}{Connection list from \code{\link[=od_get_config]{od_get_config()}} / \code{\link[=od_connection]{od_connection()}}.}
}
\value{
Character vector of branch names.
}
\description{
List branches in a project
}
\examples{
\dontrun{
od_list_branches()
}
}
\seealso{
Other repository:
\code{\link[=od_get_branch]{od_get_branch()}},
\code{\link[=od_get_commit]{od_get_commit()}},
\code{\link[=od_get_default_branch]{od_get_default_branch()}},
\code{\link[=od_get_file]{od_get_file()}},
\code{\link[=od_get_file_text]{od_get_file_text()}},
\code{\link[=od_get_tag]{od_get_tag()}},
\code{\link[=od_list_tags]{od_list_tags()}},
\code{\link[=od_query_commits]{od_query_commits()}}
}
\concept{repository}
