% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_requests.R
\name{od_request_pull_request_changes}
\alias{od_request_pull_request_changes}
\title{Request changes on a pull request}
\usage{
od_request_pull_request_changes(pull_request_number, conn = NULL)
}
\arguments{
\item{pull_request_number}{UI number.}

\item{conn}{Connection list.}
}
\value{
Parsed API response (may be \code{NULL}).
}
\description{
Request changes on a pull request
}
